ww3_diffraction: add regression tests for semi-inifnite breakwater and non-linera shoaling

+++ model/src/w3gdatmd.F90
@@ -752,9 +752,9 @@ MODULE W3GDATMD
     INTEGER*2, POINTER   :: IOBP_LOC(:)
     INTEGER*1, POINTER   :: IOBDP_LOC(:)
     INTEGER*1, POINTER   :: IOBPA_LOC(:)
-    REAL*8, POINTER      :: DIFRM(:)
-    REAL*8, POINTER      :: DIFRX(:)
-    REAL*8, POINTER      :: DIFRY(:)
+    REAL, POINTER      :: DIFRM(:)
+    REAL, POINTER      :: DIFRX(:)
+    REAL, POINTER      :: DIFRY(:)
 #endif
 
     REAL(8), POINTER      :: LEN(:,:),SI(:), IEN(:,:)
@@ -1116,9 +1116,9 @@ MODULE W3GDATMD
   INTEGER*2, POINTER   :: IOBP_LOC(:)
   INTEGER*1, POINTER   :: IOBDP_LOC(:)
   INTEGER*1, POINTER   :: IOBPA_LOC(:)
-  REAL*8, POINTER        :: DIFRM(:)
-  REAL*8, POINTER        :: DIFRX(:)
-  REAL*8, POINTER        :: DIFRY(:)
+  REAL, POINTER        :: DIFRM(:)
+  REAL, POINTER        :: DIFRX(:)
+  REAL, POINTER        :: DIFRY(:)
 #endif
 
   REAL(8), POINTER        :: IEN(:,:), LEN(:,:), SI(:)
